admin: remove admin service
