+++ demos/counter/pkg/testutils/cluster/cluster.go
@@ -145,8 +145,6 @@ func (c *Cluster) addFuddleNode() error {
 	conf := config.DefaultConfig()
 	conf.BindRegistryAddr = rpcLn.Addr().String()
 	conf.AdvRegistryAddr = rpcLn.Addr().String()
-	conf.BindAdminAddr = adminLn.Addr().String()
-	conf.AdvAdminAddr = adminLn.Addr().String()
 	conf.Locality = "us-east-1-a"
 
 	s := fuddle.New(
